+++ realm/realm-library/src/androidTest/java/io/realm/TestHelper.java
@@ -55,9 +55,9 @@ import io.realm.entities.PrimaryKeyAsBoxedInteger;
 import io.realm.entities.PrimaryKeyAsBoxedLong;
 import io.realm.entities.PrimaryKeyAsBoxedShort;
 import io.realm.entities.PrimaryKeyAsString;
-import io.realm.internal.Collection;
+import io.realm.internal.OsResults;
 import io.realm.internal.OsObject;
-import io.realm.internal.SharedRealm;
+import io.realm.internal.OsSharedRealm;
 import io.realm.internal.Table;
 import io.realm.internal.async.RealmThreadPoolExecutor;
 import io.realm.log.LogLevel;
@@ -214,29 +214,29 @@ public class TestHelper {
     /**
      * Creates an empty table whose name is "temp" with 1 column of all our supported column types, currently 7 columns.
      *
-     * @param sharedRealm A {@link SharedRealm} where the table is created.
+     * @param osSharedRealm A {@link OsSharedRealm} where the table is created.
      * @return created table.
      */
-    public static Table createTableWithAllColumnTypes(SharedRealm sharedRealm) {
-        return createTableWithAllColumnTypes(sharedRealm, "temp");
+    public static Table createTableWithAllColumnTypes(OsSharedRealm osSharedRealm) {
+        return createTableWithAllColumnTypes(osSharedRealm, "temp");
     }
 
     /**
      * Creates an empty table with 1 column of all our supported column types, currently 7 columns.
      *
-     * @param sharedRealm A {@link SharedRealm} where the table is created.
+     * @param osSharedRealm A {@link OsSharedRealm} where the table is created.
      * @param name name of the table.
      * @return created table.
      */
     @SuppressWarnings("WeakerAccess")
-    public static Table createTableWithAllColumnTypes(SharedRealm sharedRealm,
+    public static Table createTableWithAllColumnTypes(OsSharedRealm osSharedRealm,
             @SuppressWarnings("SameParameterValue") String name) {
-        boolean wasInTransaction = sharedRealm.isInTransaction();
+        boolean wasInTransaction = osSharedRealm.isInTransaction();
         if (!wasInTransaction) {
-            sharedRealm.beginTransaction();
+            osSharedRealm.beginTransaction();
         }
         try {
-            Table t = sharedRealm.createTable(name);
+            Table t = osSharedRealm.createTable(name);
 
             t.addColumn(RealmFieldType.BINARY, "binary");
             t.addColumn(RealmFieldType.BOOLEAN, "boolean");
@@ -249,43 +249,43 @@ public class TestHelper {
             return t;
         } catch (RuntimeException e) {
             if (!wasInTransaction) {
-                sharedRealm.cancelTransaction();
+                osSharedRealm.cancelTransaction();
             }
             throw e;
         } finally {
-            if (!wasInTransaction && sharedRealm.isInTransaction()) {
-                sharedRealm.commitTransaction();
+            if (!wasInTransaction && osSharedRealm.isInTransaction()) {
+                osSharedRealm.commitTransaction();
             }
         }
     }
 
-    public static Table createTable(SharedRealm sharedRealm, String name) {
-        return createTable(sharedRealm, name, null);
+    public static Table createTable(OsSharedRealm osSharedRealm, String name) {
+        return createTable(osSharedRealm, name, null);
     }
 
     public interface AdditionalTableSetup {
         void execute(Table table);
     }
 
-    public static Table createTable(SharedRealm sharedRealm, String name, AdditionalTableSetup additionalSetup) {
-        boolean wasInTransaction = sharedRealm.isInTransaction();
+    public static Table createTable(OsSharedRealm osSharedRealm, String name, AdditionalTableSetup additionalSetup) {
+        boolean wasInTransaction = osSharedRealm.isInTransaction();
         if (!wasInTransaction) {
-            sharedRealm.beginTransaction();
+            osSharedRealm.beginTransaction();
         }
         try {
-            Table table = sharedRealm.createTable(name);
+            Table table = osSharedRealm.createTable(name);
             if (additionalSetup != null) {
                 additionalSetup.execute(table);
             }
             return table;
         } catch (RuntimeException e) {
             if (!wasInTransaction) {
-                sharedRealm.cancelTransaction();
+                osSharedRealm.cancelTransaction();
             }
             throw e;
         } finally {
-            if (!wasInTransaction && sharedRealm.isInTransaction()) {
-                sharedRealm.commitTransaction();
+            if (!wasInTransaction && osSharedRealm.isInTransaction()) {
+                osSharedRealm.commitTransaction();
             }
         }
     }
@@ -1008,19 +1008,19 @@ public class TestHelper {
      * This helper method is useful to create a mocked {@link RealmResults}.
      *
      * @param realm a {@link Realm} or a {@link DynamicRealm} instance.
-     * @param collection a {@link Collection} instance.
+     * @param osResults a {@link OsResults} instance.
      * @param tableClass a Class of Table.
      * @return a created {@link RealmResults} instance.
      */
     public static <T extends RealmObject> RealmResults<T> newRealmResults(
-            BaseRealm realm, Collection collection, Class<T> tableClass) {
+            BaseRealm realm, OsResults osResults, Class<T> tableClass) {
         //noinspection TryWithIdenticalCatches
         try {
             final Constructor<RealmResults> c = RealmResults.class.getDeclaredConstructor(
-                    BaseRealm.class, Collection.class, Class.class);
+                    BaseRealm.class, OsResults.class, Class.class);
             c.setAccessible(true);
             //noinspection unchecked
-            return c.newInstance(realm, collection, tableClass);
+            return c.newInstance(realm, osResults, tableClass);
         } catch (NoSuchMethodException e) {
             throw new RuntimeException(e);
         } catch (InstantiationException e) {